+++ lib/pgslice/cli/unprep.rb
@@ -9,9 +9,14 @@ module PgSlice
       assert_table(intermediate_table)
 
       queries = [
-        "DROP TABLE #{quote_table(intermediate_table)} CASCADE;",
-        "DROP FUNCTION IF EXISTS #{quote_ident(trigger_name)}();"
+        "DROP TABLE #{quote_table(intermediate_table)} CASCADE;"
       ]
+
+      _, _, _, _, declarative, _ = intermediate_table.fetch_settings(table.trigger_name)
+      unless declarative
+        queries << "DROP FUNCTION IF EXISTS #{quote_ident(trigger_name)}();"
+      end
+
       run_queries(queries)
     end
   end