+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.7.1 (unreleased)
 
-- Improved `analyze` for declarative partitioning
+- Improved `analyze` and `unprep` for declarative partitioning
 
 ## 0.7.0 (2025-05-26)